feat: improve hTokenBalance script
